+++ src/environments/dsw.ts
@@ -14,11 +14,8 @@ export const dsw = {
         ],
         bgColorClasses: ['', 'cl1', 'cl2', 'cl3', 'cl4', 'cl5', 'cl6', 'cl7', 'cl8', 'cl9'],
         fontColors: ['fc1', 'fc2', 'fc3', 'fc4', 'fc5'],
-        icons: ['', '\uf0e4', '\uf114', '\uf080', '\uf1fe', '\uf200', '\uf201',
-            '\uf153', '\uf155', '\uf158', '\uf0c5', '\uf03a', '\uf0ce', '\uf0d1',
-            '\uf007', '\uf183', '\uf0c0', '\uf0b0', '\uf1c0', '\uf1b2', '\uf1b3',
-            '\uf02d', '\uf073', '\uf0ac', '\uf005', '\uf071', '\uf05a',
-            '\uf104'],
+        icons: ['', 'widget-1.svg', 'folder-1.svg', 'dashboard-1.svg', 'back-1.svg'],
+
         timeout: 60000,
         ver: pkg.version,
         emptyWidgetClass: 'MDX2JSON.EmptyPortlet'.toLowerCase()